+++ changelog.md
@@ -1,10 +1,15 @@
-### 0.4.0
-* RMQ queue shuffling support
+### 0.4.1 (2017.05.30)
+* RMQ queue bucketing added
+* Shuffling removed from RMQ FIFO queue
+
+### 0.4.0 (2017.05.29)
+* Caching requests with errors (PageCache API changed)
+* RMQ queue shuffling support, experimental (removed in 0.4.1)
 * Chaining RMQ requests instead of paralleling
 * JsonLinesFeedExporter now appends new lines instead of rewriting
 * FileSystemPageCache paralleled
 * Download errors logging severity decreased
 * Proper URIs merging in DownloadingActor
 
-### 0.3.0
+### 0.3.0 (2017.05.26)
 * Baseline release